requirements.txt
httpx[http2]==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1
rich==13.7.1
click==8.1.7
aiosqlite==0.20.0
dnspython==2.6.1
pydantic==2.7.1
beautifulsoup4==4.12.3
lxml==5.2.1
jinja2==3.1.4
python-whois==0.9.4
tldextract==5.1.2
yarl==1.9.4
colorama==0.4.6
anyio==4.3.0
